+++ src/extension.js
@@ -129,7 +129,6 @@ export default class JustPerfection extends Extension
             {
                 MessageTray,
                 Main,
-                GLib,
                 Gio,
                 Settings: settings,
             },